% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilikoi.featuresSelection.r
\name{lilikoi.featuresSelection}
\alias{lilikoi.featuresSelection}
\title{A featuresSelection Function}
\usage{
lilikoi.featuresSelection(PDSmatrix, threshold = 0.5, method = "info")
}
\arguments{
\item{PDSmatrix}{from PDSfun function}

\item{threshold}{to select the top pathways}

\item{method}{information gain ("info") or gain ratio ("gain")}
}
\value{
A list of top metabolites or pathways.
}
\description{
This function allows you to reduce the pathway diemsion using xxxx
}
\examples{
\donttest{
dt <- Loaddata(file=system.file("extdata", "plasma_breast_cancer.csv", package = "lilikoi2"))
Metadata <- dt$Metadata
dataSet <- dt$dataSet
Metabolite_pathway_table=MetaTOpathway('name')
PDSmatrix= PDSfun(Metabolite_pathway_table)
selected_Pathways_Weka= lilikoi.featuresSelection(PDSmatrix,threshold= 0.54,method="gain")
}


}
\keyword{features}
\keyword{selection}
