% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilikoi.Loaddata.r
\name{lilikoi.Loaddata}
\alias{lilikoi.Loaddata}
\title{A Loaddata Function}
\usage{
lilikoi.Loaddata(filename)
}
\arguments{
\item{filename}{file name.}
}
\value{
A data frame named Metadata.
}
\description{
This function allows you to load your metabolomics data.
}
\examples{
\donttest{
 Loaddata(file=system.file("extdata", "plasma_breast_cancer.csv", package = "lilikoi"))
}
}
\keyword{Load}
