% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilikoi.machine_learning.r
\name{lilikoi.machine_learning}
\alias{lilikoi.machine_learning}
\title{A machine learning Function}
\usage{
lilikoi.machine_learning(
  MLmatrix = PDSmatrix,
  measurementLabels = Label,
  significantPathways = selected_Pathways_Weka,
  trainportion = 0.8,
  cvnum = 10
)
}
\arguments{
\item{MLmatrix}{selected pathway deregulation score or metabolites expression matrix}

\item{measurementLabels}{measurement label for samples}

\item{significantPathways}{selected pathway names}

\item{trainportion}{train percentage of the total sample size}

\item{cvnum}{number of folds}
}
\value{
Evaluation results and plots of all 8 machine learning algorithms, along with variable importance plots.
}
\description{
This function for classification using 8 different machine learning algorithms
and it plots the ROC curves and the AUC, SEN, and specificty
}
\examples{
\donttest{
 measurementLabels = Metadata$Label
 MLmatrix = Metadata
 significantPathways = 0
 lilikoi.machine_learning(tPDSmatrix, measurementLabels,significantPathways,trainportion = 0.8,
    cvnum = 10)
}


}
\keyword{classifcation}
