% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilikoi.meta_path.R
\name{lilikoi.meta_path}
\alias{lilikoi.meta_path}
\title{Metabolite-pathway regression}
\usage{
lilikoi.meta_path(
  PDSmatrix,
  selected_Pathways_Weka,
  Metabolite_pathway_table,
  pathway = "Alanine, Aspartate And Glutamate Metabolism"
)
}
\arguments{
\item{PDSmatrix}{Pathway deregulation score matrix}

\item{selected_Pathways_Weka}{Selected top pathways from the featureSelection function}

\item{Metabolite_pathway_table}{Metabolites mapping table}

\item{pathway}{interested pathway name}
}
\value{
A bipartite graph of the relationships between pathways and their corresponding metabolites.
}
\description{
Performs single variate linear regression between selected pathways and each of their metabolites.
Output the network plot between pathways and metabolties.
}
