% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilikoi.preproc_knn.R
\name{lilikoi.preproc_knn}
\alias{lilikoi.preproc_knn}
\title{An imputation function.}
\usage{
lilikoi.preproc_knn(inputdata = Metadata, method = c("knn"))
}
\arguments{
\item{inputdata}{An expression data frame with samples in the rows, metabolites in the columns}

\item{method}{The method to be used to process data, including}
}
\value{
A KNN imputed dataset with samples in the rows, metabolites in the columns.
}
\description{
This function is used to preprocess data via knn imputation.
}
\examples{
\donttest{
dt <- lilikoi.Loaddata(file=system.file("extdata",
  "plasma_breast_cancer.csv", package = "lilikoi"))
Metadata <- dt$Metadata
dataSet <- dt$dataSet
lilikoi.preproc_knn(inputdata=Metadata, method="knn")
}
}
