% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilikoi.PDSfun.r
\name{lilikoi.PDSfun}
\alias{lilikoi.PDSfun}
\title{A PDSfun Function}
\usage{
lilikoi.PDSfun(qvec)
}
\arguments{
\item{qvec}{This is the Metabolite_pathway_table from MetaTOpathway function. This table includes the metabolites ids and the its corssponding hmdb ids}
}
\value{
A large matrix of the pathway deregulation scores for each pathway in different samples.
}
\description{
This function allows you to compute Pathway Desregulation Score deriving
make sure that you have the below database for the metabolites and pathway list:
meta_path.RData
}
\examples{
\donttest{
dt <- Loaddata(file=system.file("extdata", "plasma_breast_cancer.csv", package = "lilikoi2"))
Metadata <- dt$Metadata
dataSet <- dt$dataSet
convertResults=MetaTOpathway('name')
Metabolite_pathway_table = convertResults$table
PDSmatrix= lilikoi.PDSfun(Metabolite_pathway_table)
}
}
\keyword{PDS}
